refactor: common yaml test
